# services/communication-service/openapi.yaml
openapi: 3.0.3
info:
  title: WorkHub Communication Service API
  version: 1.1.0
  description: API endpoints for managing Web Push subscriptions, notifications, and preferences.
servers:
  - url: http://localhost:3002
    description: Local development server
paths:
...
  /api/notifications:
    get:
      summary: List user notifications
      description: Returns a paginated list of inbox notifications.
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: unreadOnly
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Notifications feed retrieved successfully.
...
  /api/notifications/preferences:
    get:
      summary: Get user notification preferences
      responses:
        '200':
          description: Preference settings retrieved.
    put:
      summary: Update user notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                notifyEmail:
                  type: boolean
                notifyPush:
                  type: boolean
                preferredLang:
                  type: string
                timezone:
                  type: string
      responses:
        '200':
          description: Preference settings updated.